Use activejob to process multiple users at the same time Even after the timeout corrections, with users posting videos and pictures, sometimes it takes a while for a user to be processed. With this change we can have multiple workers to make the processing faster.
