+++ daemons/check_for_statuses.rb
@@ -1,11 +1,6 @@
 require ENV["RAILS_ENV_PATH"]
-require 'mastodon_ext'
-require 'mastodon_user_processor'
-require 'twitter_user_processor'
-require 'toot_transformer'
 require 'httparty'
 require 'interruptible_sleep'
-require 'stats'
 
 class CheckForToots
   OLDER_THAN_IN_SECONDS = 30
@@ -21,19 +16,15 @@ class CheckForToots
     @@sleeper ||= InterruptibleSleep.new
   end
 
-  def self.stats
-    @@stats ||= Stats.new
-  end
-
   def self.available_since_last_check
     loop do
-      u = User.where('(posting_from_mastodon = ? OR posting_from_twitter = ?) AND (mastodon_last_check < now() - interval \'? seconds\' or twitter_last_check < now() - interval \'? seconds\')', true, true, OLDER_THAN_IN_SECONDS, OLDER_THAN_IN_SECONDS).order(mastodon_last_check: :asc, twitter_last_check: :asc).first
+      u = User.where('locked = ? AND (posting_from_mastodon = ? OR posting_from_twitter = ?) AND (mastodon_last_check < now() - interval \'? seconds\' or twitter_last_check < now() - interval \'? seconds\')', false, true, true, OLDER_THAN_IN_SECONDS, OLDER_THAN_IN_SECONDS).order(mastodon_last_check: :asc, twitter_last_check: :asc).first
       if u.nil?
         Rails.logger.debug { "No user to look at. Sleeping for #{SLEEP_FOR} seconds" }
         sleeper.sleep(SLEEP_FOR)
       else
-        stats.time('mastodon.processing_time') { MastodonUserProcessor::process_user(u) } if u.posting_from_mastodon
-        stats.time('twitter.processing_time') { TwitterUserProcessor::process_user(u) } if u.posting_from_twitter
+        u.locked = true; u.save
+        ProcessUserJob.perform_later(u.id)
       end
       break if finished
     end
@@ -50,19 +41,6 @@ Signal.trap("INT") {
 }
 
 Rails.logger.debug { "Starting" }
-
-twitter_client = Twitter::REST::Client.new do |config|
-  config.consumer_key = ENV['TWITTER_CLIENT_ID']
-  config.consumer_secret = ENV['TWITTER_CLIENT_SECRET']
-end
-
-begin
-  TootTransformer::twitter_short_url_length = twitter_client.configuration.short_url_length
-  TootTransformer::twitter_short_url_length_https = twitter_client.configuration.short_url_length_https
-rescue Twitter::Error::Forbidden
-  Rails.logger.error { "Missing Twitter credentials" }
-  exit
-end
 
 
 CheckForToots::available_since_last_check
