+++ db/migrate/20180412164151_add_locked_to_users.rb
@@ -1,0 +1,5 @@
+class AddLockedToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :locked, :boolean, default: false, null: false
+  end
+end
